+++ src/templatespecification/src/infrastructure/mod.rs
@@ -1,2 +1,0 @@
-pub mod configuration_loader;
-pub mod folder_loader;